add a benchmark using Oleg's delimcc library
Delimcc is an implementation of delimited control using a runtime hack
that is not claimed to be efficient. The bytecode implementation is
simpler/faster than the native implementation; on this benchmark,
because control costs dominate, this means that the bytecode program
is simpler/faster than the native program.

(To install, `opam install delimcc`.)

On my machine:

```
$ ocamlfind ocamlc -package delimcc -linkpkg delimcc_nondet.ml -o delimcc_nondet.byte
$ time ocamlrun ./delimcc_nondet.byte 12
14200

real	0m7.053s
user	0m7.026s
sys	0m0.007s

$ ocamlfind ocamlopt -package delimcc -linkpkg delimcc_nondet.ml -o delimcc_nondet.native
$ ./delimcc_nondet.native 12
14200

real	0m9.221s
user	0m9.155s
sys	0m0.007s
```

(in comparison, delim_nondet.ml takes about 6.7 seconds on the same input.)

This delimcc version works well with OCaml 4.04.2 and 4.05.0, but not
with the newest version 4.06.0. This may be caused by a change in the
OCaml runtime that lacks the matching change in delimcc. I will report
the issue to Oleg.