add a Prolog solution (tested with SWI-Prolog)
It is too hard to setup a Curry system (according to my experiments
with nqueens, KICS2 is very slow on search problems and PAKS requires
an existing Prolog setup, making it strictly harder to install than
a Prolog), so I'm adding a native Prolog version which we may use for
comparison instead.

Tested with SWI-prolog. The `prolog.sh script` gives an interface
compatible with `run_benchmark_time.rb`:

```
$ ruby ./run_benchmark_time.rb "sh prolog.sh"
sh prolog.sh
4: 89.0
5: 92.2
6: 91.4
7: 92.6
8: 115.4
9: 195.6
10: 637.4
11: 3086.8
12: 18689.4
```