OCaml benchmarks: consistently precompute (range n)
The replay-based implementations are paying a large performance cost
to recompute (range n) on each call to the nqueens function, as this
is part of the pure computation that gets duplicate on each replay.

Nqueens should be a case where pure computation costs are dominated by
effective computations (no pun intended), namely the exploration of
the large search space. Factorizing list allocations allows to better
reflect the performance characteristics of various control-operator
approaches.