OCaml benchmark: filter okay placements on the fly
I found tihs optimization in "Eff Directly in OCaml",

  Oleg Kiselyov and KC Sivaramakrishnan, 2017
  http://kcsrk.info/papers/caml-eff17.pdf

and it has a large impact on the performance of the "effect" version,
but also speeds up all other benchmarks by quite a lot -- it
significantly decreases the branching factor at each level, and
removes the use of `fail ()`.

I liked the previous code better, but I think that it is important to
write the fastest possible code that is not cheating. Otherwise their
paper has a surprisingly fast Effect implementation, and our work has
a surprisingly slower Effect implementation, and it looks bad.

TODO: reflect the changes in the SML versions, and re-run the benchmarks.