build delimcc from source rather than OPAM
The 2017 version of delimcc doesn't work with GCC 7.x, which means
that some evaluators may get random build failures (... and
myself). The 2018 version of delimcc fixes this issue (I reported the
bug to Oleg), but it is not packaged in OPAM yet, so here we just
build it from source.